+++ dataworkspace/dataworkspace/tests/test_admin.py
@@ -2247,6 +2247,44 @@ class TestSourceLinkAdmin(BaseAdminTestCase):
 
 
 class TestDatasetAdmin(BaseAdminTestCase):
+    def test_edit_dataset_authorized_users(self):
+        dataset = factories.DataSetFactory.create()
+        user1 = factories.UserFactory.create()
+        user2 = factories.UserFactory.create()
+        factories.DataSetUserPermissionFactory.create(dataset=dataset, user=user1)
+
+        self.assertEqual(dataset.user_has_access(user1), True)
+        self.assertEqual(dataset.user_has_access(user2), False)
+
+        response = self._authenticated_post(
+            reverse('admin:datasets_datacutdataset_change', args=(dataset.id,)),
+            {
+                'published': dataset.published,
+                'name': dataset.name,
+                'slug': dataset.slug,
+                'short_description': 'test short description',
+                'description': 'test description',
+                'type': 2,
+                'sourcelink_set-TOTAL_FORMS': '0',
+                'sourcelink_set-INITIAL_FORMS': '0',
+                'sourcelink_set-MIN_NUM_FORMS': '0',
+                'sourcelink_set-MAX_NUM_FORMS': '1000',
+                'sourceview_set-TOTAL_FORMS': '0',
+                'sourceview_set-INITIAL_FORMS': '0',
+                'sourceview_set-MIN_NUM_FORMS': '0',
+                'sourceview_set-MAX_NUM_FORMS': '1000',
+                'customdatasetquery_set-TOTAL_FORMS': '0',
+                'customdatasetquery_set-INITIAL_FORMS': '0',
+                'customdatasetquery_set-MIN_NUM_FORMS': '0',
+                'customdatasetquery_set-MAX_NUM_FORMS': '1000',
+                'authorized_users': user2.id,
+                '_continue': 'Save and continue editing',
+            },
+        )
+        self.assertContains(response, 'was changed successfully')
+        self.assertEqual(dataset.user_has_access(user1), False)
+        self.assertEqual(dataset.user_has_access(user2), True)
+
     def test_delete_external_source_link(self):
         dataset = factories.DataSetFactory.create()
         source_link = factories.SourceLinkFactory(